refactor(experimental): rename agent http.py to client.py for parity with inf
inf-cli puts the same kind of HTTP client wrappers in
cli/inference/client.py; rename the agent counterpart to match so the
two sibling subcommand packages have parallel layouts. Pure rename +
three import updates, no behavior change.
